Optimized Content Browser — It's not updating if not hovered. Added Folder-History buttons. Folders now renderer first. Content browser now supports cyrillic. Small ImGui UI fixes. Added new Utils function to convert wstring to u8string.

+++ Eagle/src/Eagle/Classes/StaticMesh.cpp
@@ -142,7 +142,7 @@ namespace Eagle
 
 		std::vector<StaticMesh> meshes;
 		processNode(scene->mRootNode, scene, meshes, filename, bLazy);
-		int meshesCount = meshes.size();
+		size_t meshesCount = meshes.size();
 
 		if (meshesCount == 0)
 			return MakeRef<StaticMesh>();
